+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Set `initialClickTarget` to `null` when element gets destroyed
+- Avoid closing the dialog element when there are nested autocomplete elements which are still open
 
 ## [0.1.3] - 2023-08-03